src/requirements.txt
jinja2==2.10.1
mysqlclient==1.4.2.post1
django==2.2.2
django-reversion
pytz
django-recaptcha
social-auth-app-django
pyyaml
Pillow